Remove ContentDialogWindow layer, major refactoring
