+++ WinUI3Example/ExampleDialog.xaml
@@ -1,5 +1,5 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<essential:ContentDialogWindow
+<?xml version="1.0" encoding="utf-8" ?>
+<essential:WindowedContentDialog
     x:Class="WinUI3Example.ExampleDialog"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -10,7 +10,4 @@
     Title="ExampleDialog"
     mc:Ignorable="d">
 
-    <!--<essential:ContentDialogWindow.DialogContent>
-        <TextBlock Text="This is a dialog" />
-    </essential:ContentDialogWindow.DialogContent>-->
-</essential:ContentDialogWindow>
+</essential:WindowedContentDialog>
