--- WinUI3Package/ContentDialogWindow.idl
@@ -1,35 +1,0 @@
-﻿
-import "ContentDialogWindowButtonClickEventArgs.idl";
-import "ContentDialogContent.idl";
-
-namespace WinUI3Package
-{
-    [default_interface]
-    unsealed runtimeclass ContentDialogWindow : Microsoft.UI.Xaml.Window
-    {
-        ContentDialogWindow();
-
-        void Open();
-
-        void Hide();
-
-        void UpdateWindowSize();
-
-        void SetParent(Microsoft.UI.Xaml.Window parent, Boolean center);
-
-        Windows.Foundation.IAsyncOperation<Microsoft.UI.Xaml.Controls.ContentDialogResult> ShowAsync();
-
-        Microsoft.UI.Xaml.Media.Imaging.BitmapImage HeaderImage;
-        String HeaderImageUri;
-
-        Microsoft.UI.Xaml.Controls.ContentDialogResult Result{ get; };
-
-        Boolean HasTitleBar;
-        Boolean IsResizable;
-
-        event Windows.Foundation.TypedEventHandler<WinUI3Package.ContentDialogWindow, Object> Loaded;
-        event Windows.Foundation.TypedEventHandler<WinUI3Package.ContentDialogWindow, Object> Opened;
-
-        event Windows.Foundation.TypedEventHandler<Object, Microsoft.UI.Xaml.Input.RightTappedRoutedEventArgs> ContentRightTapped;
-    }
-}
