+++ WinUI3Package/WindowedContentDialog.xaml
@@ -1,32 +1,24 @@
-<?xml version="1.0" encoding="utf-8" ?>
+﻿<?xml version="1.0" encoding="utf-8" ?>
 <Window
-    x:Class="WinUI3Package.ContentDialogWindow"
+    x:Class="WinUI3Package.WindowedContentDialog"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:local="using:WinUI3Package"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    Title="ContentDialogWindow"
+    Title="WindowedContentDialog"
     mc:Ignorable="d">
 
     <Grid x:Name="ContentDialogGrid">
-
-
         <StackPanel>
-
             <Grid
                 x:Name="ImageGrid"
                 Width="0"
                 Height="0"
                 Visibility="Collapsed">
-
                 <Image x:Name="WindowImage" />
-
             </Grid>
-
-            <local:ContentDialogContent
-                x:Name="ContentDialogContentControl"
-                Loaded="OnContentLoaded" />
+            <local:ContentDialogContent x:Name="ContentDialogContentControl" Loaded="OnContentLoaded" />
         </StackPanel>
 
     </Grid>
